First Commit of the site
Haven't changed any of the files. There is quite an amount of extraneous
files that have no business being downloaded. Including: the reset from
yahoo, foundation file can be optimized as 95% not used, normalize not
used, stellar js and css not even implemented, as well as all the css
and js files should be concatenated except those served by cms.