First editing of the site itself
added sass to write the css easier so install scout to edit and play
with it. also added a single clicked image to be displayed behind a
discog item placed so the user doesnt have to download an entire image
for every item clicked. also working on reducing the amount of css and
js as you have to download around 1 mb worth of data each site refress
and thats horrible for mobile.