Completely integrated nav bar and Discog section
In this version I have completely implemented the Nav bar & About
section except fonts; as well as the discography section with the
clickable album covers. Took a little bit to implement due to the many
extraneous CSS files which I had to completely read and understand to
see if it was worth ridding of. I unimplemented quite a few css and js
files and the site seems surprisingly functional. I foresee only a few
more days until the gist of this site is complete at least in my eyes.